+++ webknossos/tests/dataset/test_dataset.py
@@ -67,8 +67,8 @@ OUTPUT_PATHS = [TESTOUTPUT_DIR, REMOTE_TESTOUTPUT_DIR]
 
 
 def copy_simple_dataset(
-    data_format: DataFormat, output_path: Path, suffix: str | None = None
-) -> Path:
+    data_format: DataFormat, output_path: UPath, suffix: str | None = None
+) -> UPath:
     suffix = (f"_{suffix}") if suffix is not None else ""
     new_dataset_path = output_path / f"simple_{data_format}_dataset{suffix}"
     rmtree(new_dataset_path)
@@ -80,8 +80,8 @@ def copy_simple_dataset(
 
 
 def prepare_dataset_path(
-    data_format: DataFormat, output_path: Path, suffix: str | None = None
-) -> Path:
+    data_format: DataFormat, output_path: UPath, suffix: str | None = None
+) -> UPath:
     suffix = (f"_{suffix}") if suffix is not None else ""
     new_dataset_path = output_path / f"{data_format}_dataset{suffix}"
     rmtree(new_dataset_path)
@@ -208,7 +208,7 @@ def assure_exported_properties(ds: Dataset) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_create_dataset_with_layer_and_mag(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
 
@@ -235,7 +235,7 @@ def test_create_dataset_with_layer_and_mag(
 
 
 @pytest.mark.parametrize("output_path", [TESTOUTPUT_DIR, REMOTE_TESTOUTPUT_DIR])
-def test_ome_ngff_0_4_metadata(output_path: Path) -> None:
+def test_ome_ngff_0_4_metadata(output_path: UPath) -> None:
     ds_path = prepare_dataset_path(DataFormat.Zarr, output_path)
     ds = Dataset(ds_path, voxel_size=(11, 11, 28))
     layer = ds.add_layer("color", COLOR_CATEGORY, data_format=DataFormat.Zarr)
@@ -278,7 +278,7 @@ def test_ome_ngff_0_4_metadata(output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("output_path", [TESTOUTPUT_DIR, REMOTE_TESTOUTPUT_DIR])
-def test_ome_ngff_0_5_metadata(output_path: Path) -> None:
+def test_ome_ngff_0_5_metadata(output_path: UPath) -> None:
     ds_path = prepare_dataset_path(DataFormat.Zarr3, output_path)
     ds = Dataset(ds_path, voxel_size=(11, 11, 28))
     layer = ds.add_layer("color", COLOR_CATEGORY, data_format=DataFormat.Zarr3)
@@ -472,7 +472,7 @@ def test_deprecated_chunks_per_shard() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_open_dataset(data_format: DataFormat, output_path: Path) -> None:
+def test_open_dataset(data_format: DataFormat, output_path: UPath) -> None:
     new_dataset_path = copy_simple_dataset(data_format, output_path)
     ds = Dataset.open(new_dataset_path)
 
@@ -482,7 +482,7 @@ def test_open_dataset(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_modify_existing_dataset(data_format: DataFormat, output_path: Path) -> None:
+def test_modify_existing_dataset(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds1 = Dataset(ds_path, voxel_size=(1, 1, 1))
     ds1.add_layer(
@@ -510,7 +510,7 @@ def test_modify_existing_dataset(data_format: DataFormat, output_path: Path) -> 
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_view_read(data_format: DataFormat, output_path: Path) -> None:
+def test_view_read(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path)
 
     with pytest.warns(UserWarning, match=".*not aligned with the shard shape.*"):
@@ -528,7 +528,7 @@ def test_view_read(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_view_write(data_format: DataFormat, output_path: Path) -> None:
+def test_view_write(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path)
     with pytest.warns(UserWarning, match=".*not aligned with the shard shape.*"):
         wk_view = (
@@ -551,7 +551,7 @@ def test_view_write(data_format: DataFormat, output_path: Path) -> None:
 
 @pytest.mark.parametrize("output_path", [TESTOUTPUT_DIR, REMOTE_TESTOUTPUT_DIR])
 @pytest.mark.parametrize("data_format", [DataFormat.Zarr, DataFormat.Zarr3])
-def test_direct_zarr_access(output_path: Path, data_format: DataFormat) -> None:
+def test_direct_zarr_access(output_path: UPath, data_format: DataFormat) -> None:
     ds_path = copy_simple_dataset(data_format, output_path)
     mag = Dataset.open(ds_path).get_layer("color").get_mag("1")
 
@@ -571,7 +571,7 @@ def test_direct_zarr_access(output_path: Path, data_format: DataFormat) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_view_write_out_of_bounds(data_format: DataFormat, output_path: Path) -> None:
+def test_view_write_out_of_bounds(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(
         data_format, output_path, "view_dataset_out_of_bounds"
     )
@@ -592,7 +592,7 @@ def test_view_write_out_of_bounds(data_format: DataFormat, output_path: Path) ->
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_mag_view_write_out_of_bounds(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "dataset_out_of_bounds")
 
@@ -612,7 +612,7 @@ def test_mag_view_write_out_of_bounds(
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_mag_view_write_out_of_bounds_mag2(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "dataset_out_of_bounds")
 
@@ -635,7 +635,7 @@ def test_mag_view_write_out_of_bounds_mag2(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_view_write_allow_resize(data_format: DataFormat, output_path: Path) -> None:
+def test_view_write_allow_resize(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     layer = Dataset(ds_path, voxel_size=(1, 1, 1)).add_layer("color", COLOR_CATEGORY)
     mag = layer.add_mag("1")
@@ -673,7 +673,9 @@ def test_view_write_allow_resize(data_format: DataFormat, output_path: Path) -> 
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_view_write_allow_unaligned(data_format: DataFormat, output_path: Path) -> None:
+def test_view_write_allow_unaligned(
+    data_format: DataFormat, output_path: UPath
+) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     layer = Dataset(ds_path, voxel_size=(1, 1, 1)).add_layer(
         "color",
@@ -720,7 +722,7 @@ def test_view_write_allow_unaligned(data_format: DataFormat, output_path: Path) 
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_views_are_equal(data_format: DataFormat, output_path: Path) -> None:
+def test_views_are_equal(data_format: DataFormat, output_path: UPath) -> None:
     np.random.seed(1234)
     data: np.ndarray = (np.random.rand(10, 10, 10) * 255).astype(np.uint8)
 
@@ -758,7 +760,7 @@ def test_views_are_equal(data_format: DataFormat, output_path: Path) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_update_new_bounding_box_offset(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -819,7 +821,7 @@ def test_chunked_compressed_write() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_write_multi_channel_uint8(data_format: DataFormat, output_path: Path) -> None:
+def test_write_multi_channel_uint8(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "multichannel")
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
     mag = ds.add_layer(
@@ -837,7 +839,7 @@ def test_write_multi_channel_uint8(data_format: DataFormat, output_path: Path) -
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_wkw_write_multi_channel_uint16(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "multichannel")
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -860,7 +862,7 @@ def test_wkw_write_multi_channel_uint16(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_empty_read(data_format: DataFormat, output_path: Path) -> None:
+def test_empty_read(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "empty")
     mag = (
         Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -875,7 +877,7 @@ def test_empty_read(data_format: DataFormat, output_path: Path) -> None:
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 @pytest.mark.parametrize("absolute_offset", [None, Vec3Int(12, 12, 12)])
 def test_write_layer(
-    data_format: DataFormat, output_path: Path, absolute_offset: Vec3Int | None
+    data_format: DataFormat, output_path: UPath, absolute_offset: Vec3Int | None
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "empty")
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -901,7 +903,7 @@ def test_write_layer(
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 @pytest.mark.parametrize("absolute_offset", [None, Vec3Int(12, 12, 12)])
 def test_write_layer_mag2(
-    data_format: DataFormat, output_path: Path, absolute_offset: Vec3Int | None
+    data_format: DataFormat, output_path: UPath, absolute_offset: Vec3Int | None
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "empty")
     ds = Dataset(ds_path, voxel_size=(12, 12, 24))
@@ -933,7 +935,7 @@ def test_write_layer_mag2(
 @pytest.mark.parametrize("absolute_offset", [None, (3, 12, 12, 12)])
 def test_write_layer_5d(
     data_format: DataFormat,
-    output_path: Path,
+    output_path: UPath,
     absolute_offset: VecIntLike | None,
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "empty")
@@ -958,7 +960,7 @@ def test_write_layer_5d(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_read_padded_data(data_format: DataFormat, output_path: Path) -> None:
+def test_read_padded_data(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "empty")
     mag = (
         Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -976,7 +978,7 @@ def test_read_padded_data(data_format: DataFormat, output_path: Path) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_num_channel_mismatch_assertion(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -996,7 +998,7 @@ def test_num_channel_mismatch_assertion(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_get_or_add_layer(data_format: DataFormat, output_path: Path) -> None:
+def test_get_or_add_layer(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
 
@@ -1041,7 +1043,7 @@ def test_get_or_add_layer(data_format: DataFormat, output_path: Path) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_get_or_add_layer_idempotence(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -1056,7 +1058,7 @@ def test_get_or_add_layer_idempotence(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_get_or_add_mag(data_format: DataFormat, output_path: Path) -> None:
+def test_get_or_add_mag(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
 
     layer = Dataset(ds_path, voxel_size=(1, 1, 1)).add_layer(
@@ -1209,7 +1211,7 @@ def test_properties_with_segmentation() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_relative_mag_paths(data_format: DataFormat, output_path: Path) -> None:
+def test_relative_mag_paths(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path)
 
     ds = Dataset.open(ds_path)
@@ -1229,7 +1231,7 @@ def test_relative_mag_paths(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_chunking_wk(data_format: DataFormat, output_path: Path) -> None:
+def test_chunking_wk(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(2, 2, 1))
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -1294,7 +1296,7 @@ def test_chunking_wkw_advanced(data_format: DataFormat) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_chunking_wkw_wrong_chunk_shape(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(
         data_format, output_path, "chunking_with_wrong_chunk_shape"
@@ -1369,7 +1371,7 @@ def test_dataset_exist_ok() -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_changing_layer_bounding_box(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "change_layer_bounding_box")
     ds = Dataset.open(ds_path)
@@ -1433,7 +1435,7 @@ def test_changing_layer_bounding_box(
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_dataset_bounding_box_calculation(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "change_layer_bounding_box")
     ds = Dataset.open(ds_path)
@@ -1581,7 +1583,7 @@ def test_adding_layer_with_valid_dtype_per_layer() -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_writing_subset_of_compressed_data_multi_channel(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "compressed_data")
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -1626,7 +1628,7 @@ def test_writing_subset_of_compressed_data_multi_channel(
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_writing_subset_of_compressed_data_single_channel(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "compressed_data")
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -1672,7 +1674,7 @@ def test_writing_subset_of_compressed_data_single_channel(
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_writing_subset_of_compressed_data(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "compressed_data")
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -1734,7 +1736,7 @@ def test_writing_subset_of_compressed_data(
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_writing_subset_of_chunked_compressed_data(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "compressed_data")
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -1995,7 +1997,7 @@ def test_remote_add_symlink_mag(data_format: DataFormat) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_add_copy_mag(data_format: DataFormat, output_path: Path) -> None:
+def test_add_copy_mag(data_format: DataFormat, output_path: UPath) -> None:
     original_ds_path = prepare_dataset_path(data_format, output_path, "original")
     copy_ds_path = prepare_dataset_path(data_format, output_path, "copy")
 
@@ -2043,7 +2045,7 @@ def test_add_copy_mag(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_add_fs_copy_mag(data_format: DataFormat, output_path: Path) -> None:
+def test_add_fs_copy_mag(data_format: DataFormat, output_path: UPath) -> None:
     original_ds_path = prepare_dataset_path(data_format, output_path, "original")
     copy_ds_path = prepare_dataset_path(data_format, output_path, "copy")
 
@@ -2093,7 +2095,7 @@ def test_add_fs_copy_mag(data_format: DataFormat, output_path: Path) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_search_dataset_also_for_long_layer_name(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "long_layer_name")
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
@@ -2312,7 +2314,7 @@ def test_dataset_conversion_wkw_only() -> None:
 @pytest.mark.parametrize("output_path", [TESTOUTPUT_DIR, REMOTE_TESTOUTPUT_DIR])
 @pytest.mark.parametrize("data_format", [DataFormat.Zarr, DataFormat.Zarr3])
 def test_dataset_conversion_from_wkw_to_zarr(
-    output_path: Path, data_format: DataFormat
+    output_path: UPath, data_format: DataFormat
 ) -> None:
     converted_path = prepare_dataset_path(data_format, output_path, "converted")
 
@@ -2408,7 +2410,7 @@ def _func_invalid_target_chunk_shape_wk(args: tuple[View, View, int]) -> None:
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
 def test_for_zipped_chunks_invalid_target_chunk_shape_wk(
-    data_format: DataFormat, output_path: Path
+    data_format: DataFormat, output_path: UPath
 ) -> None:
     ds_path = prepare_dataset_path(
         data_format, output_path, "zipped_chunking_source_invalid"
@@ -2453,7 +2455,7 @@ def test_for_zipped_chunks_invalid_target_chunk_shape_wk(
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_read_only_view(data_format: DataFormat, output_path: Path) -> None:
+def test_read_only_view(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "read_only_view")
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
     mag = ds.get_or_add_layer(
@@ -2478,7 +2480,7 @@ def test_read_only_view(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_bounding_box_on_disk(data_format: DataFormat, output_path: Path) -> None:
+def test_bounding_box_on_disk(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(2, 2, 1))
     chunk_shape, shard_shape = default_chunk_config(data_format, 8)
@@ -2545,7 +2547,7 @@ def test_bounding_box_on_disk(data_format: DataFormat, output_path: Path) -> Non
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_compression(data_format: DataFormat, output_path: Path) -> None:
+def test_compression(data_format: DataFormat, output_path: UPath) -> None:
     new_dataset_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(new_dataset_path, voxel_size=(2, 2, 1))
     mag1 = ds.add_layer(
@@ -2807,7 +2809,7 @@ def test_read_bbox() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_add_copy_layer(data_format: DataFormat, output_path: Path) -> None:
+def test_add_copy_layer(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path, "original")
     copy_path = prepare_dataset_path(data_format, output_path, "copy")
 
@@ -2862,7 +2864,7 @@ def test_add_copy_layer(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_rename_layer(data_format: DataFormat, output_path: Path) -> None:
+def test_rename_layer(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
     layer = ds.add_layer("color", COLOR_CATEGORY, data_format=data_format)
@@ -2901,7 +2903,7 @@ def test_rename_layer(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_delete_layer_and_mag(data_format: DataFormat, output_path: Path) -> None:
+def test_delete_layer_and_mag(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
     color_layer = ds.add_layer("color", COLOR_CATEGORY, data_format=data_format)
@@ -2957,7 +2959,7 @@ def test_delete_layer_and_mag(data_format: DataFormat, output_path: Path) -> Non
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_add_layer_like(data_format: DataFormat, output_path: Path) -> None:
+def test_add_layer_like(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds = Dataset(ds_path, voxel_size=(1, 1, 1))
     color_layer1 = ds.add_layer(
@@ -3101,7 +3103,7 @@ def test_pickle_view() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_warn_outdated_properties(data_format: DataFormat, output_path: Path) -> None:
+def test_warn_outdated_properties(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = prepare_dataset_path(data_format, output_path)
     ds1 = Dataset(ds_path, voxel_size=(1, 1, 1))
     ds2 = Dataset.open(ds_path)
@@ -3161,7 +3163,7 @@ def test_can_compress_mag8() -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_downsampling(data_format: DataFormat, output_path: Path) -> None:
+def test_downsampling(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "downsampling")
 
     color_layer = Dataset.open(ds_path).get_layer("color")
@@ -3184,7 +3186,7 @@ def test_downsampling(data_format: DataFormat, output_path: Path) -> None:
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_aligned_downsampling(data_format: DataFormat, output_path: Path) -> None:
+def test_aligned_downsampling(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "aligned_downsampling")
     dataset = Dataset.open(ds_path)
     input_layer = dataset.get_layer("color")
@@ -3229,7 +3231,7 @@ def test_aligned_downsampling(data_format: DataFormat, output_path: Path) -> Non
 
 
 @pytest.mark.parametrize("data_format,output_path", DATA_FORMATS_AND_OUTPUT_PATHS)
-def test_guided_downsampling(data_format: DataFormat, output_path: Path) -> None:
+def test_guided_downsampling(data_format: DataFormat, output_path: UPath) -> None:
     ds_path = copy_simple_dataset(data_format, output_path, "guided_downsampling")
 
     input_dataset = Dataset.open(ds_path)
@@ -3308,7 +3310,9 @@ def test_zarr_copy_to_remote_dataset(data_format: DataFormat) -> None:
 
 @pytest.mark.parametrize("input_path", OUTPUT_PATHS)
 @pytest.mark.parametrize("output_path", OUTPUT_PATHS)
-def test_fs_copy_dataset_with_attachments(input_path: Path, output_path: Path) -> None:
+def test_fs_copy_dataset_with_attachments(
+    input_path: UPath, output_path: UPath
+) -> None:
     ds_path = copy_simple_dataset(DEFAULT_DATA_FORMAT, input_path)
     new_ds_path = prepare_dataset_path(DEFAULT_DATA_FORMAT, output_path, "copied")