+++ webknossos/webknossos/dataset/mag_view.py
@@ -110,7 +110,7 @@ class MagView(View):
         layer: "Layer",
         mag: Mag,
         *,
-        path: Path,
+        path: UPath,
         chunk_shape: Vec3Int,
         shard_shape: Vec3Int | None = None,
         chunks_per_shard: Vec3Int | None = None,
@@ -157,7 +157,7 @@ class MagView(View):
         self,
         layer: "Layer",
         mag: Mag,
-        path: Path,
+        path: UPath,
         read_only: bool = False,
     ) -> None:
         """
@@ -209,11 +209,11 @@ class MagView(View):
         return self._layer
 
     @property
-    def path(self) -> Path:
+    def path(self) -> UPath:
         """Get the path to this magnification level's data.
 
         Returns:
-            Path: Path to the data files on disk.
+            UPath: Path to the data files on disk.
 
         Notes:
             - Path may be local or remote depending on dataset configuration